chore: Renaming emoji picker file. #275
